refactor(management commands): Creates a new CommandUtils class with useful tools
I grow weary of the repetition and near-repetition here. Let's start centralizing.
