+++ cl/lib/command_utils.py
@@ -1,5 +1,7 @@
 import logging
-from django.core.management import BaseCommand
+import os
+
+from django.core.management import BaseCommand, CommandError
 
 logger = logging.getLogger(__name__)
 
@@ -13,3 +15,15 @@ class VerboseCommand(BaseCommand):
             logger.setLevel(logging.INFO)
         elif verbosity > 1:
             logger.setLevel(logging.DEBUG)
+
+
+class CommandUtils(object):
+    """A mixin to give some useful methods to sub classes."""
+
+    @staticmethod
+    def ensure_file_ok(file_path):
+        """Check to make sure that a file path exists and is valid."""
+        if not os.path.exists(file_path):
+            raise CommandError("Unable to find file at %s" % file_path)
+        if not os.access(file_path, os.R_OK):
+            raise CommandError("Unable to read file at %s" % file_path)
